modulo int64_t int32_t

+++ include/structure/math/modulo.h
@@ -18,8 +18,8 @@ void modulo_normalize_int(I* v, const I& M) {
 	if (*v < 0 || *v >= M) *v %= M;
 	if (*v < 0) *v += M;
 }
-inline void modulo_normalize(long long* v, long long M) { modulo_normalize_int(v, M); }
-inline void modulo_normalize(int* v, int M) { modulo_normalize_int(v, M); }
+inline void modulo_normalize(int64_t* v, int64_t M) { modulo_normalize_int(v, M); }
+inline void modulo_normalize(int32_t* v, int32_t M) { modulo_normalize_int(v, M); }
 
 // modulo multiplication
 template<typename T>
@@ -36,16 +36,16 @@ I modulo_long_multiply_int(I x, I y, I M) {
 	}
 	return r;
 }
-inline long long modulo_multiply(long long x, long long y, long long M) {
+inline int64_t modulo_multiply(int64_t x, int64_t y, int64_t M) {
 	modulo_normalize_int(&x, M);
 	modulo_normalize_int(&y, M);
 	bool fit = (x < (1LL << 31) && y < (1LL << 31));
 	return fit ? (x * y) % M : modulo_long_multiply_int(x, y, M);
 }
-inline int modulo_multiply(int x, int y, int M) {
+inline int32_t modulo_multiply(int32_t x, int32_t y, int32_t M) {
 	modulo_normalize_int(&x, M);
 	modulo_normalize_int(&y, M);
-	return ((long long)x * y) % M;
+	return ((int64_t)x * y) % M;
 }
 
 // modulo inversion
@@ -61,8 +61,8 @@ template<typename I> I modulo_inverse_int(I v, I M) {
 	modulo_normalize_int(&vi, M);
 	return vi;
 }
-inline long long modulo_inverse(long long v, long long M) { return modulo_inverse_int(v, M); }
-inline int modulo_inverse(int v, int M) { return modulo_inverse_int(v, M); }
+inline int64_t modulo_inverse(int64_t v, int64_t M) { return modulo_inverse_int(v, M); }
+inline int32_t modulo_inverse(int32_t v, int32_t M) { return modulo_inverse_int(v, M); }
 
 // modulo division
 template<typename T> T modulo_divide(const T& x, const T& y, const T& M) {
@@ -85,8 +85,8 @@ I modulo_divide_int(I x, I y, I M) {
 	}
 	return modulo_multiply(x, yi, M);
 }
-inline long long modulo_divide(long long x, long long y, long long M) { return modulo_divide_int(x, y, M); }
-inline int modulo_divide(int x, int y, int M) { return modulo_divide_int(x, y, M); }
+inline int64_t modulo_divide(int64_t x, int64_t y, int64_t M) { return modulo_divide_int(x, y, M); }
+inline int32_t modulo_divide(int32_t x, int32_t y, int32_t M) { return modulo_divide_int(x, y, M); }
 
 template<typename T, int ID, bool STATIC = true>
 struct modulo_members;
